refactor: Change SnsUserPath
- Change Path to Security

#16
